#61 Code cleanup
