+++ code/FinanceManager.Components/Components/AccountDetailsPageContents/BankAccountComponents/ManageBankAccount.razor
@@ -1,4 +1,5 @@
 ﻿@using FinanceManager.Domain.Enums
+
 @if (BankAccount is not null)
 {
     <MudForm @ref="form" @bind-IsValid="@success" @bind-Errors="@errors">
