+++ code/FinanceManager.Components/Components/AccountDetailsPageContents/BankAccountComponents/RemoveBankEntry.razor
@@ -2,6 +2,7 @@
 @using FinanceManager.Domain.Entities.Accounts.Entries
 @using FinanceManager.Domain.Enums
 @using FinanceManager.Domain.Repositories
+
 @if (BankAccountEntry is not null)
 {
     <MudContainer MaxWidth="MaxWidth.Medium">
@@ -17,12 +18,15 @@
             <MudItem xs="12" sm="6">
                 <MudNumericField Required="true" HideSpinButtons="true" @bind-Value="BalanceChange" Label="Balance change" Variant="Variant.Text" AdornmentText="@_currency" Adornment="Adornment.End" ReadOnly />
             </MudItem>
+
             <MudItem xs="12" sm="6">
                 <MudDatePicker Label="Posting date" @bind-Date="date" ReadOnly />
             </MudItem>
+
             <MudItem xs="12" sm="6">
                 <MudTextField @bind-Value="BankAccountEntry.Description" T="string" Label="Description" Variant="Variant.Text" MaxLines="5" AutoGrow="true" ReadOnly />
             </MudItem>
+
             <MudItem xs="12">
                 <MudButton Variant="Variant.Filled" Color="Color.Primary" DropShadow="false" OnClick=Cancel>Cancel</MudButton>
                 <MudButton Variant="Variant.Outlined" Color="Color.Secondary" DropShadow="false" OnClick=Confirm Class="mx-2">Confirm</MudButton>
@@ -37,20 +41,12 @@
     private string ExpenseType = string.Empty;
     private string _currency = "PLN";
 
-    [Parameter]
-    public required Func<Task> Confirm { get; set; }
-
-    [Parameter]
-    public required Func<Task> Cancel { get; set; }
-
-    [Parameter]
-    public required string BankAccountName { get; set; }
-
-    [Parameter]
-    public required BankAccountEntry BankAccountEntry { get; set; }
-
-    [Inject]
-    public required ISettingsService SettingsService { get; set; }
+    [Parameter] public required Func<Task> Confirm { get; set; }
+    [Parameter] public required Func<Task> Cancel { get; set; }
+    [Parameter] public required string BankAccountName { get; set; }
+    [Parameter] public required BankAccountEntry BankAccountEntry { get; set; }
+    
+    [Inject] public required ISettingsService SettingsService { get; set; }
 
     protected override void OnParametersSet()
     {
