+++ code/FinanceManager.Components/Components/Dashboard/Cards/TimeSeries/AssetsTimeSeriesCard.razor
@@ -1,11 +1,5 @@
 ﻿@using ApexCharts
-@using FinanceManager.Domain.Entities.Accounts
-@using FinanceManager.Domain.Entities.MoneyFlowModels
-@using FinanceManager.Domain.Providers
-@using FinanceManager.Domain.Repositories
-@using FinanceManager.Domain.Repositories.Account
-@using FinanceManager.Domain.Services
-@using FinanceManager.Components.ViewModels
+@using FinanceManager.Domain.Entities.MoneyFlowModels;
 
 <div class="card">
     <div class="row m-0">
@@ -14,9 +8,9 @@
         </div>
     </div>
     <div class="row" style="min-height: 150px; position:relative; bottom:-2px">
-        <ApexChart @ref="chart" TItem="TimeSeriesModel" Height="150" Options=@options XAxisType="XAxisType.Datetime">
+        <ApexChart @ref="_chart" TItem="TimeSeriesModel" Height="150" Options=@_options XAxisType="XAxisType.Datetime">
             <ApexPointSeries TItem="TimeSeriesModel"
-            Items="priceTimeseries"
+            Items="_priceTimeseries"
             Name="Assets value"
             XValue="e => e.DateTime.ToLocalTime()"
             YValue="e=> e.Value"
