+++ code/FinanceManager.Components/Components/Dashboard/Cards/TimeSeries/InvestmentTypeTimeSeriesCard.razor
@@ -1,22 +1,13 @@
 ﻿@using ApexCharts
-@using FinanceManager.Domain.Entities.Accounts
-@using FinanceManager.Domain.Entities.MoneyFlowModels
-@using FinanceManager.Domain.Providers
-@using FinanceManager.Domain.Repositories
-@using FinanceManager.Domain.Repositories.Account
-@using FinanceManager.Domain.Services
-@using FinanceManager.Components.Helpers
-@using FinanceManager.Components.ViewModels
-@using Microsoft.Extensions.Logging
-
+@using FinanceManager.Domain.Entities.MoneyFlowModels;
 
 <div class="card">
     <div class="card-body">
         <h5 class="card-title">Assets per type</h5>
         <div class="row" style="min-height: 150px">
-            <ApexChart @ref="chart" TItem="TimeSeriesModel" Height="150" Options=@options XAxisType="XAxisType.Datetime">
+            <ApexChart @ref="_chart" TItem="TimeSeriesModel" Height="150" Options=@_options XAxisType="XAxisType.Datetime">
                 <ApexPointSeries TItem="TimeSeriesModel"
-                                 Items="priceTimeseries"
+                                 Items="_priceTimeseries"
                                  Name="Assets value"
                                  XValue="e => e.DateTime.ToLocalTime()"
                                  YValue="e=> e.Value"
