+++ code/FinanceManager.Components/_Imports.razor
@@ -2,5 +2,5 @@
 @using FinanceManager.Domain.Services
 @using MudBlazor
 @using Microsoft.AspNetCore.Components.Authorization
-
+@using Microsoft.AspNetCore.Components;
 @using FinanceManager.Components.Services;
