+++ code/FinanceManager/Pages/Assets.razor
@@ -1,6 +1,5 @@
 ﻿@page "/Assets"
-@using FinanceManager.Components.Components.Dashboard
-@using FinanceManager.Components.Components.Dashboard.Cards
+
 
 <PageTitle>Assets</PageTitle>
