+++ code/FinanceManager/Pages/Investments.razor
@@ -10,7 +10,7 @@
             <div class="card">
                 <div class="card-body">
                     <h5 class="card-title">Investments value over time - Time series</h5>
-                    <ChartSpinner></ChartSpinner>
+                    <ChartSpinner/>
                 </div>
             </div>
         </div>
@@ -28,7 +28,7 @@
             <div class="card">
                 <div class="card-body">
                     <h5 class="card-title">Repayment and forecast - Time series</h5>
-                    <ChartSpinner></ChartSpinner>
+                    <ChartSpinner/>
                 </div>
             </div>
         </div>
@@ -37,7 +37,7 @@
             <div class="card">
                 <div class="card-body">
                     <h5 class="card-title">Real estate - Time series</h5>
-                    <ChartSpinner></ChartSpinner>
+                    <ChartSpinner/>
                 </div>
             </div>
         </div>
@@ -46,13 +46,9 @@
             <div class="card">
                 <div class="card-body">
                     <h5 class="card-title">Crypto - Time series</h5>
-                    <ChartSpinner></ChartSpinner>
+                    <ChartSpinner/>
                 </div>
             </div>
         </div>
     </div>
 </div>
-
-@code {
-
-}
