+++ code/FinanceManager/_Imports.razor
@@ -20,3 +20,7 @@
 @using System.Collections.ObjectModel
 
 @using MudBlazor
+
+@using FinanceManager.Components.Components.Dashboard
+@using FinanceManager.Components.Components.Dashboard.Cards
+@using FinanceManager.Components.Components.Dashboard.Cards.TimeSeries
